Cleaning up the index interface a little, so we can extend the row indexer from the base KVIndex
